+++ src/main/java/com/github/fireduck64/sockthing/sharesaver/ShareSaver.java
@@ -1,5 +1,7 @@
 package com.github.fireduck64.sockthing.sharesaver;
 
+import java.math.BigInteger;
+
 import com.github.fireduck64.sockthing.PoolUser;
 import com.github.fireduck64.sockthing.SubmitResult;
 
@@ -12,6 +14,6 @@ public interface ShareSaver
      * The unique_job_string is what this share unit will be deduped on.  It must be unique for each valid submit.
      */
     public void saveShare(PoolUser pu, SubmitResult submitResult, String source, String uniqueJobString,
-                          long blockReward, long feeTotal)
+                          BigInteger blockReward, BigInteger feeTotal)
     throws ShareSaveException;
 }
